refactor: inline string encoding rules
